Refactoring code is end!

+++ src/hpp/order_manager.hpp
@@ -16,12 +16,12 @@ namespace ORDERMANAGER
 	class OrderItem
 	{
 	public:
-		OrderItem(const uint16_t& _order_id)
+		OrderItem(const uint32_t& _order_id)
 		{
 			this->item.order_id = _order_id;
 		}
 
-		void addMeal(const uint16_t &meal_id, const uint8_t &quantity);
+		void addMeal(const uint32_t &meal_id, const uint8_t &quantity);
 
 		void displayOrder();
 	private:
@@ -29,10 +29,10 @@ namespace ORDERMANAGER
 		{
 		public:
 			/* The ID must be unique: */
-			uint16_t order_id{};
+			uint32_t order_id{};
 
 			/* uint16_t = meal-id | uint8_t = quantity */
-			std::map <uint16_t, uint8_t> meal_quantity {};
+			std::map <uint32_t, uint8_t> meal_quantity {};
 		};
 
 		Item item {};
@@ -46,10 +46,10 @@ namespace ORDERMANAGER
 	class Order
 	{
 	public:
-		void createOrder(const uint16_t &order_id);
-		OrderItem& getItem(const uint16_t order_id);
+		void createOrder(const uint32_t &order_id);
+		OrderItem& getItem(const uint32_t order_id);
 	private:
-		static std::map <uint16_t, OrderItem> current_orders;
+		static std::map <uint32_t, OrderItem> current_orders;
 	};
 	/*------------------------------------------------------------------------------------------------------------------------*/
